requirements.txt
langchain
langchain_community
torch
transformers
ctransformers
sentence_transformers
openai
faiss_cpu
streamlit
huggingface_hub
python-dotenv
pypdf
pdfminer.six
requests
tiktoken
accelerate
bitsandbytes
einops
bs4
chromadb
safetensors
pandas
